fixed on dispatcher:
-create database if not exist
-sort
-date format
-button's style
-date validation
-offer's number
-offer generator
-add address on offer details

+++ src/main/java/com/truckcompany/service/WaybillService.java
@@ -151,7 +151,7 @@ public class WaybillService {
             .map(g -> {
                 Goods goods = new Goods();
                 goods.setName(g.getName());
-                goods.setUncheckedNumber(g.getCount());
+                goods.setUncheckedNumber((long)g.getCount());
                 goods.setType(g.getType());
                 goods.setPrice(g.getPrice());
                 goods.setState("UNCHECKED");
